CommonHandler: emit compile-time warning when an annotated class (CObj, Cxx) doesn't explicitly inherit CObject or CxxObject
Reason: we're using the type hierarchy during macro expansion to determine if object's passed to external calls are wrappers. We're adding CObject / CxxObject as parent to every annotated class during expansion, but depending on the compilation sequence, a CObj/Cxx class might not yet have been expanded when we're checking if it is a wrapper => the user must ensure that all wrapper explicitly inherit from CObject.