Cxx: emit inline source annotations separately on classes and companion objects
Previously, we emitted the InlineSource annotation always on the class or trait.
This resulted in the C++ wrapper snippets not being generated, if the program
didn't actually use the type (i.e. the class/trait) but only called static
methods in the companion object.

with this change we always emit the InlineSource annotation for methods defined in
an object on the object itself.